Pullup ticket #2739 - requested by he parrot: build fix
Revisions pulled up:
- lang/parrot/distinfo				1.8
- lang/parrot/patches/patch-ad			1.3
- lang/parrot/patches/patch-ae			1.1
---
Module Name:	pkgsrc
Committed By:	he
Date:		Tue Apr 14 23:50:20 UTC 2009

Modified Files:
	pkgsrc/lang/parrot: distinfo
Added Files:
	pkgsrc/lang/parrot/patches: patch-ad patch-ae

Log Message:
Fix the detection and use of perldoc, so that this package can be
fully built and installed when the build is run as root.  The reason
this is problematical is that perldoc changes uid if run as root, to
either 'nobody', 'nouser' or uid -2, for security reasons, and then
cannot write files owned by root or create files in directories owned
by root with 'normal' permissions.

No revision bump as this should purely be a build fix.